agent_builder/requirements.txt
# Python requirements for MAAP Agent Builder

# Core dependencies
flask>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
uuid>=1.30
tiktoken>=0.5.1
pipe21>=0.2.0

# LangChain & related packages
langchain>=0.0.340
langchain-core>=0.1.5
langchain-community>=0.0.15
langchain-openai>=0.0.5
langchain-anthropic>=0.0.5
langchain-aws>=0.0.2
langchain-fireworks>=0.1.1
langchain-together>=0.0.2
langchain-cohere>=0.0.1
langchain-huggingface>=0.0.7
langchain-mongodb>=0.0.2
langchain-voyageai>=0.0.1

# LangGraph
langgraph>=0.0.20
langgraph-checkpoint-mongodb


# MongoDB dependencies
pymongo>=4.5.0
certifi>=2023.7.22

# Vector databases & embedding models
sentence-transformers>=2.2.2

# API & model provider dependencies
openai>=1.3.0
anthropic>=0.5.0
cohere>=4.32
boto3
fireworks-ai>=0.7.0
voyageai>=0.1.0

# Web server
gunicorn>=21.2.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1

# Optional: MCP integration
# mcp>=0.1.0
langchain-mcp-adapters>=0.1.0
